backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.4
google-generativeai==0.3.2
elevenlabs==0.2.27
PyPDF2==3.0.1
reportlab==4.0.9
PyJWT==2.8.0
cryptography==42.0.2
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
python-multipart==0.0.6
